requirements.txt
Django==2.0.4
djangorestframework==3.8.2
gunicorn==19.7.1
psycopg2==2.7.4
pytz==2018.4
requests==2.18.4
Pillow==5.1.0
whitenoise==3.3.1